requirements.txt
flask
flask-auth
sqlalchemy
pyjwt
bcrypt
flask_cors
